requirements.txt
numpy
tensorflow==1.8.0
matplotlib
sklearn
scipy
captcha
pillow